+++ packages/go_router/example/lib/books/src/widgets/author_list.dart
@@ -9,11 +9,7 @@ import '../data.dart';
 /// The author list view.
 class AuthorList extends StatelessWidget {
   /// Creates an [AuthorList].
-  const AuthorList({
-    required this.authors,
-    this.onTap,
-    super.key,
-  });
+  const AuthorList({required this.authors, this.onTap, super.key});
 
   /// The list of authors to be shown.
   final List<Author> authors;
@@ -23,15 +19,12 @@ class AuthorList extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => ListView.builder(
-        itemCount: authors.length,
-        itemBuilder: (BuildContext context, int index) => ListTile(
-          title: Text(
-            authors[index].name,
-          ),
-          subtitle: Text(
-            '${authors[index].books.length} books',
-          ),
+    itemCount: authors.length,
+    itemBuilder:
+        (BuildContext context, int index) => ListTile(
+          title: Text(authors[index].name),
+          subtitle: Text('${authors[index].books.length} books'),
           onTap: onTap != null ? () => onTap!(authors[index]) : null,
         ),
-      );
+  );
 }