+++ packages/interactive_media_ads/test/android/ads_manager_test.dart
@@ -69,9 +69,9 @@ void main() {
         mockAdsRenderingSettings.setLoadVideoTimeout(2000),
         mockAdsRenderingSettings.setMimeTypes(<String>['value']),
         mockAdsRenderingSettings.setPlayAdsAfterTime(5.0),
-        mockAdsRenderingSettings.setUiElements(
-          <ima.UiElement>[ima.UiElement.countdown],
-        ),
+        mockAdsRenderingSettings.setUiElements(<ima.UiElement>[
+          ima.UiElement.countdown,
+        ]),
         mockAdsRenderingSettings.setEnableCustomTabs(true),
         mockAdsManager.init(mockAdsRenderingSettings),
       ]);
@@ -120,17 +120,12 @@ void main() {
     test('onAdEvent', () async {
       final MockAdsManager mockAdsManager = MockAdsManager();
 
-      late final void Function(
-        ima.AdEventListener,
-        ima.AdEvent,
-      ) onAdEventCallback;
+      late final void Function(ima.AdEventListener, ima.AdEvent)
+      onAdEventCallback;
 
       final InteractiveMediaAdsProxy proxy = InteractiveMediaAdsProxy(
         newAdEventListener: ({
-          required void Function(
-            ima.AdEventListener,
-            ima.AdEvent,
-          ) onAdEvent,
+          required void Function(ima.AdEventListener, ima.AdEvent) onAdEvent,
         }) {
           onAdEventCallback = onAdEvent;
           return MockAdEventListener();
@@ -164,20 +159,16 @@ void main() {
     test('onAdErrorEvent', () async {
       final MockAdsManager mockAdsManager = MockAdsManager();
 
-      late final void Function(
-        ima.AdErrorListener,
-        ima.AdErrorEvent,
-      ) onAdErrorCallback;
+      late final void Function(ima.AdErrorListener, ima.AdErrorEvent)
+      onAdErrorCallback;
 
       final InteractiveMediaAdsProxy proxy = InteractiveMediaAdsProxy(
         newAdEventListener: ({required dynamic onAdEvent}) {
           return MockAdEventListener();
         },
         newAdErrorListener: ({
-          required void Function(
-            ima.AdErrorListener,
-            ima.AdErrorEvent,
-          ) onAdError,
+          required void Function(ima.AdErrorListener, ima.AdErrorEvent)
+          onAdError,
         }) {
           onAdErrorCallback = onAdError;
           return MockAdErrorListener();
@@ -199,8 +190,9 @@ void main() {
       final MockAdErrorEvent mockErrorEvent = MockAdErrorEvent();
       final MockAdError mockError = MockAdError();
       when(mockError.errorType).thenReturn(ima.AdErrorType.load);
-      when(mockError.errorCode)
-          .thenReturn(ima.AdErrorCode.adsRequestNetworkError);
+      when(
+        mockError.errorCode,
+      ).thenReturn(ima.AdErrorCode.adsRequestNetworkError);
       when(mockError.message).thenReturn('error message');
       when(mockErrorEvent.error).thenReturn(mockError);
       onAdErrorCallback(MockAdErrorListener(), mockErrorEvent);