+++ packages/pigeon/pigeons/proxy_api_tests.dart
@@ -4,11 +4,7 @@
 
 import 'package:pigeon/pigeon.dart';
 
-enum ProxyApiTestEnum {
-  one,
-  two,
-  three,
-}
+enum ProxyApiTestEnum { one, two, three }
 
 /// The core ProxyApi test class that each supported host language must
 /// implement in platform_tests integration tests.
@@ -115,23 +111,25 @@ abstract class ProxyApiTestClass extends ProxyApiSuperClass
   /// Returns the passed list with ProxyApis, to test serialization and
   /// deserialization.
   late List<ProxyApiTestClass?> Function(List<ProxyApiTestClass?> aList)
-      flutterEchoProxyApiList;
+  flutterEchoProxyApiList;
 
   /// Returns the passed map, to test serialization and deserialization.
   late Map<String?, Object?> Function(Map<String?, Object?> aMap)
-      flutterEchoMap;
+  flutterEchoMap;
 
   /// Returns the passed map with ProxyApis, to test serialization and
   /// deserialization.
   late Map<String?, ProxyApiTestClass?> Function(
-      Map<String?, ProxyApiTestClass?> aMap) flutterEchoProxyApiMap;
+    Map<String?, ProxyApiTestClass?> aMap,
+  )
+  flutterEchoProxyApiMap;
 
   /// Returns the passed enum to test serialization and deserialization.
   late ProxyApiTestEnum Function(ProxyApiTestEnum anEnum) flutterEchoEnum;
 
   /// Returns the passed ProxyApi to test serialization and deserialization.
   late ProxyApiSuperClass Function(ProxyApiSuperClass aProxyApi)
-      flutterEchoProxyApi;
+  flutterEchoProxyApi;
 
   // ========== Nullable argument/return type tests ==========
 
@@ -155,15 +153,15 @@ abstract class ProxyApiTestClass extends ProxyApiSuperClass
 
   /// Returns the passed map, to test serialization and deserialization.
   late Map<String?, Object?>? Function(Map<String?, Object?>? aMap)?
-      flutterEchoNullableMap;
+  flutterEchoNullableMap;
 
   /// Returns the passed enum to test serialization and deserialization.
   late ProxyApiTestEnum? Function(ProxyApiTestEnum? anEnum)?
-      flutterEchoNullableEnum;
+  flutterEchoNullableEnum;
 
   /// Returns the passed ProxyApi to test serialization and deserialization.
   late ProxyApiSuperClass? Function(ProxyApiSuperClass? aProxyApi)?
-      flutterEchoNullableProxyApi;
+  flutterEchoNullableProxyApi;
 
   // ========== Async tests ==========
   // These are minimal since async FlutterApi only changes Dart generation.
@@ -217,9 +215,7 @@ abstract class ProxyApiTestClass extends ProxyApiSuperClass
 
   /// Returns the passed list with ProxyApis, to test serialization and
   /// deserialization.
-  List<ProxyApiTestClass> echoProxyApiList(
-    List<ProxyApiTestClass> aList,
-  );
+  List<ProxyApiTestClass> echoProxyApiList(List<ProxyApiTestClass> aList);
 
   /// Returns the passed map, to test serialization and deserialization.
   Map<String?, Object?> echoMap(Map<String?, Object?> aMap);
@@ -403,14 +399,16 @@ abstract class ProxyApiTestClass extends ProxyApiSuperClass
 
   @async
   List<ProxyApiTestClass?> callFlutterEchoProxyApiList(
-      List<ProxyApiTestClass?> aList);
+    List<ProxyApiTestClass?> aList,
+  );
 
   @async
   Map<String?, Object?> callFlutterEchoMap(Map<String?, Object?> aMap);
 
   @async
   Map<String?, ProxyApiTestClass?> callFlutterEchoProxyApiMap(
-      Map<String?, ProxyApiTestClass?> aMap);
+    Map<String?, ProxyApiTestClass?> aMap,
+  );
 
   @async
   ProxyApiTestEnum callFlutterEchoEnum(ProxyApiTestEnum anEnum);