+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/legacy/platform_interface/webview_cookie_manager.dart
@@ -29,7 +29,8 @@ abstract class WebViewCookieManagerPlatform extends PlatformInterface {
   static set instance(WebViewCookieManagerPlatform? instance) {
     if (instance == null) {
       throw AssertionError(
-          'Platform interfaces can only be set to a non-null instance');
+        'Platform interfaces can only be set to a non-null instance',
+      );
     }
     PlatformInterface.verify(instance, _token);
     _instance = instance;
@@ -40,12 +41,14 @@ abstract class WebViewCookieManagerPlatform extends PlatformInterface {
   /// Returns true if cookies were present before clearing, else false.
   Future<bool> clearCookies() {
     throw UnimplementedError(
-        'clearCookies is not implemented on the current platform');
+      'clearCookies is not implemented on the current platform',
+    );
   }
 
   /// Sets a cookie for all [WebView] instances.
   Future<void> setCookie(WebViewCookie cookie) {
     throw UnimplementedError(
-        'setCookie is not implemented on the current platform');
+      'setCookie is not implemented on the current platform',
+    );
   }
 }